DataGetter/SysInfo: allow running without Linux::Distribution
That module is not distributed in any of the usual Arch repositories and
needs to be built from AUR. This is always a problem, since they need to
be rebuilt every time a major perl version is installed.
However, getting the info is fairly straightforward on most distros
nowadays, so just check the /etc/lsb-release file if the module is
missing.